Add state dumps to fatal errors
This adds a complete dump of the VM state to all fatal error panic
messages. Closes #20.

Generating these state dumps requires a clone, so they are disabled by
default in `eval()` for performance reasons. A boolean parameter to
`eval()` may be used to enable state dumps in fatal error panics.