Remove QuickCheck
This fixes the build.

It turns out that using QuickCheck for tests in subpackages currently won't work due to rust-lang/cargo/#860. Since Travis runs tests for SVM and Scheme as subcrates rather than as the root crate, this won't work on CI currently. Ah well. It was a beautiful dream.