Don't test subcrates on travis
Tests for subcrates should be run on the Travis build jobs for those subcrates. Since we are now getting subcrates from crates.io, we no longer need to run tests on them, as we know they have only been deployed to crates.io if all of their tests pass. This should speed up the build for this crate significantly, and allows us to include dev-dependencies in subcrates (i.e. debreaks quickcheck).

This commit was meant to go to master but was pushed to bytecode-dev by accident